fix: MainActivityの処理をListFragmentに移行
